Tier 1 + 2 cleanups (CharacterConfigStore, dialogs, menus, view model)
Behavior-preserving refactors. No logic changes except where noted.

core:
- CharacterConfigStore: the Section enum now carries its own
  get/clear/encode/decodeInto behavior, collapsing the four parallel
  when(section) blocks (adding a section is a single edit). New ConfigFileIo
  holds the shared atomic temp-file write, ensure-parent-dir, and
  template-aware encode primitives, now used by both CharacterConfigStore and
  TomlFileStore (the duplicated write path is gone).

compose:
- EditHighlightDialog: stop mutating the styles list during composition. The
  list is grown in a keyed LaunchedEffect instead; the render loop is capped at
  the available slots and the styles[0] read is guarded, so the frame before
  the effect tops up the list can't read out of bounds. Only effect: a
  newly-needed row for a brand-new highlight appears one frame later.
- ComposeTextStream.toStreamLine: extract the slow-line diagnostic into
  logSlowLine() so the function reads as render -> retry-if-slow -> log.
- GameViewModel.init: extract the eight inline flow pipelines into named
  trackX()/observeX() functions (promoted one ctor param to a val).
- DialogContent: move the ~556 lines of @Preview fixtures to a sibling
  DialogContentPreviews.kt (769 -> 212 lines).

desktopApp:
- TitleBarView/AppMenuBar: define the app menu items once as a platform-agnostic
  AppMenu model, rendered as a native menu bar on macOS and as popups elsewhere
  (was two hand-maintained copies).

Verified: all modules compile (JVM + Android); core and scripting tests pass;
ktlint clean. iOS not built locally (needs macOS); no iOS-specific code changed.
