+++ compose/src/mobileMain/kotlin/warlockfe/warlock3/compose/ui/window/DialogContent.kt
@@ -210,560 +210,3 @@ private fun DialogSkin(data: DialogObject.Skin) {
         }
     }
 }
-
-@Preview
-@Composable
-private fun VitalBarsPreview() {
-    CompositionLocalProvider(
-        LocalSkin provides
-            mapOf(
-                "HealthBar" to
-                    SkinObject(
-                        children =
-                            mapOf(
-                                "health" to
-                                    SkinObject(
-                                        color = SkinColor("#FFFFFF"),
-                                        bar = SkinColor("#800000"),
-                                        background = SkinColor("#444444"),
-                                    ),
-                            ),
-                    ),
-            ),
-    ) {
-        val dialogData =
-            listOf(
-                DialogObject.Skin(
-                    id = "healthSkin",
-                    left = null,
-                    top = null,
-                    width = null,
-                    height = null,
-                    align = null,
-                    topAnchor = null,
-                    leftAnchor = null,
-                    tooltip = null,
-                    name = "healthBar",
-                    controls = listOf("health"),
-                ),
-                DialogObject.ProgressBar(
-                    id = "health",
-                    left = DataDistance.Percent(value = Percentage(value = 0)),
-                    top = DataDistance.Percent(value = Percentage(value = 0)),
-                    width = DataDistance.Percent(value = Percentage(value = 25)),
-                    height = DataDistance.Percent(value = Percentage(value = 100)),
-                    align = null,
-                    topAnchor = null,
-                    leftAnchor = null,
-                    tooltip = null,
-                    value = Percentage(value = 50),
-                    text = "health 13/26",
-                ),
-                DialogObject.Skin(
-                    id = "manaSkin",
-                    left = DataDistance.Percent(value = Percentage(value = 25)),
-                    top = DataDistance.Percent(value = Percentage(value = 0)),
-                    width = DataDistance.Percent(value = Percentage(value = 25)),
-                    height = DataDistance.Percent(value = Percentage(value = 100)),
-                    align = null,
-                    topAnchor = null,
-                    leftAnchor = null,
-                    tooltip = null,
-                    name = "manaBar",
-                    controls = listOf("mana"),
-                ),
-                DialogObject.ProgressBar(
-                    id = "mana",
-                    left = DataDistance.Percent(value = Percentage(value = 25)),
-                    top = DataDistance.Percent(value = Percentage(value = 0)),
-                    width = DataDistance.Percent(value = Percentage(value = 25)),
-                    height = DataDistance.Percent(value = Percentage(value = 100)),
-                    align = null,
-                    topAnchor = null,
-                    leftAnchor = null,
-                    tooltip = null,
-                    value = Percentage(value = 100),
-                    text = "mana 2/2",
-                ),
-                DialogObject.Skin(
-                    id = "spiritSkin",
-                    left = DataDistance.Percent(value = Percentage(value = 75)),
-                    top = DataDistance.Percent(value = Percentage(value = 0)),
-                    width = DataDistance.Percent(value = Percentage(value = 25)),
-                    height = DataDistance.Percent(value = Percentage(value = 100)),
-                    align = null,
-                    topAnchor = null,
-                    leftAnchor = null,
-                    tooltip = null,
-                    name = "spiritBar",
-                    controls = listOf("spirit"),
-                ),
-                DialogObject.ProgressBar(
-                    id = "spirit",
-                    left = DataDistance.Percent(value = Percentage(value = 75)),
-                    top = DataDistance.Percent(value = Percentage(value = 0)),
-                    width = DataDistance.Percent(value = Percentage(value = 25)),
-                    height = DataDistance.Percent(value = Percentage(value = 100)),
-                    align = null,
-                    topAnchor = null,
-                    leftAnchor = null,
-                    tooltip = null,
-                    value = Percentage(value = 100),
-                    text = "spirit 7/7",
-                ),
-                DialogObject.Skin(
-                    id = "staminaSkin",
-                    left = DataDistance.Percent(value = Percentage(value = 50)),
-                    top = DataDistance.Percent(value = Percentage(value = 0)),
-                    width = DataDistance.Percent(value = Percentage(value = 25)),
-                    height = DataDistance.Percent(value = Percentage(value = 100)),
-                    align = null,
-                    topAnchor = null,
-                    leftAnchor = null,
-                    tooltip = null,
-                    name = "staminaBar",
-                    controls = listOf("stamina"),
-                ),
-                DialogObject.ProgressBar(
-                    id = "stamina",
-                    left = DataDistance.Percent(value = Percentage(value = 50)),
-                    top = DataDistance.Percent(value = Percentage(value = 0)),
-                    width = DataDistance.Percent(value = Percentage(value = 25)),
-                    height = DataDistance.Percent(value = Percentage(value = 100)),
-                    align = null,
-                    topAnchor = null,
-                    leftAnchor = null,
-                    tooltip = null,
-                    value = Percentage(value = 100),
-                    text = "stamina 26/26",
-                ),
-            )
-        DialogContent(
-            dataObjects = dialogData,
-            modifier = Modifier.size(400.dp, 24.dp),
-            executeCommand = {},
-            style = StyleDefinition(),
-        )
-    }
-}
-
-@Preview
-@Composable
-private fun CombatDialogPreview() {
-    val dialogData =
-        listOf(
-            DialogObject.Image(
-                id = "unsheathe",
-                left = DataDistance.Pixels(value = -50),
-                top = DataDistance.Pixels(value = 3),
-                width = DataDistance.Pixels(value = 29),
-                height = DataDistance.Pixels(value = 29),
-                align = "n",
-                topAnchor = null,
-                leftAnchor = null,
-                tooltip = "Unsheathe Weapon",
-                name = "SwordBtn",
-                cmd = "_ready weapon",
-                echo = "ready weapon",
-            ),
-            DialogObject.Image(
-                id = "readyshield",
-                left = DataDistance.Pixels(value = 3),
-                top = null,
-                width = DataDistance.Pixels(value = 29),
-                height = DataDistance.Pixels(value = 29),
-                align = null,
-                topAnchor = null,
-                leftAnchor = "unsheathe",
-                tooltip = "Ready Shield",
-                name = "ShieldBtn",
-                cmd = "_ready shield",
-                echo = "ready shield",
-            ),
-            DialogObject.Image(
-                id = "sheathe",
-                left = DataDistance.Pixels(value = 8),
-                top = null,
-                width = DataDistance.Pixels(value = 29),
-                height = DataDistance.Pixels(value = 29),
-                align = null,
-                topAnchor = null,
-                leftAnchor = "readyshield",
-                tooltip = "Sheathe Weapon",
-                name = "NoSwordBtn",
-                cmd = "_store weapon",
-                echo = "store weapon",
-            ),
-            DialogObject.Image(
-                id = "storeshield",
-                left = DataDistance.Pixels(value = 3),
-                top = null,
-                width = DataDistance.Pixels(value = 29),
-                height = DataDistance.Pixels(value = 29),
-                align = null,
-                topAnchor = null,
-                leftAnchor = "sheathe",
-                tooltip = "Store Shield",
-                name = "NoShieldBtn",
-                cmd = "_store shield",
-                echo = "store shield",
-            ),
-            DialogObject.Link(
-                id = "lnConfigure",
-                left = DataDistance.Pixels(value = 0),
-                top = DataDistance.Pixels(value = 30),
-                width = null,
-                height = null,
-                align = "n",
-                topAnchor = null,
-                leftAnchor = null,
-                tooltip = null,
-                value = "configure",
-                cmd = "_cmbtpl configure dialog",
-                echo = "configure",
-            ),
-            DialogObject.ProgressBar(
-                id = "pbarStance",
-                left = DataDistance.Pixels(value = 0),
-                top = DataDistance.Pixels(value = 51),
-                width = DataDistance.Pixels(value = 130),
-                height = DataDistance.Pixels(value = 16),
-                align = "n",
-                topAnchor = null,
-                leftAnchor = null,
-                tooltip = "Percent of stance contributing to defense",
-                value = Percentage(value = 100),
-                text = "defensive(100 %)",
-            ),
-            DialogObject.Button(
-                id = "cmdDefStance",
-                left = DataDistance.Pixels(value = 0),
-                top = DataDistance.Pixels(value = 70),
-                width = DataDistance.Pixels(value = 55),
-                height = DataDistance.Pixels(value = 20),
-                align = "nw",
-                topAnchor = null,
-                leftAnchor = null,
-                tooltip = "Assume a Defensive Stance",
-                value = "defense",
-                cmd = "_stance defensive",
-                echo = "stance defensive",
-            ),
-            DialogObject.Button(
-                id = "cmdOffStance",
-                left = DataDistance.Pixels(value = 0),
-                top = DataDistance.Pixels(value = 70),
-                width = DataDistance.Pixels(value = 50),
-                height = DataDistance.Pixels(value = 20),
-                align = "ne",
-                topAnchor = null,
-                leftAnchor = null,
-                tooltip = "Assume an Offensive Stance",
-                value = "offense",
-                cmd = "_stance offensive",
-                echo = "stance offensive",
-            ),
-            DialogObject.Button(
-                id = "cmdTarget",
-                left = DataDistance.Pixels(value = 0),
-                top = DataDistance.Pixels(value = 93),
-                width = DataDistance.Pixels(value = 55),
-                height = DataDistance.Pixels(value = 20),
-                align = "nw",
-                topAnchor = null,
-                leftAnchor = null,
-                tooltip = "Select a Random Target",
-                value = "target",
-                cmd = "target random",
-                echo = null,
-            ),
-            DialogObject.Button(
-                id = "cmdAttack",
-                left = DataDistance.Pixels(value = 0),
-                top = DataDistance.Pixels(value = 93),
-                width = DataDistance.Pixels(value = 50),
-                height = DataDistance.Pixels(value = 20),
-                align = "ne",
-                topAnchor = null,
-                leftAnchor = null,
-                tooltip = "Attack Current Target",
-                value = "attack",
-                cmd = "attack",
-                echo = null,
-            ),
-            DialogObject.Button(
-                id = "cmdHide",
-                left = DataDistance.Pixels(value = 0),
-                top = DataDistance.Pixels(value = 116),
-                width = DataDistance.Pixels(value = 55),
-                height = DataDistance.Pixels(value = 20),
-                align = "nw",
-                topAnchor = null,
-                leftAnchor = null,
-                tooltip = "Hide",
-                value = "hide",
-                cmd = "hide",
-                echo = null,
-            ),
-            DialogObject.Button(
-                id = "cmdAmbush",
-                left = DataDistance.Pixels(value = 0),
-                top = DataDistance.Pixels(value = 116),
-                width = DataDistance.Pixels(value = 50),
-                height = DataDistance.Pixels(value = 20),
-                align = "ne",
-                topAnchor = null,
-                leftAnchor = null,
-                tooltip = "Ambush Current Target",
-                value = "ambush",
-                cmd = "ambush",
-                echo = null,
-            ),
-            DialogObject.Button(
-                id = "cmdJab",
-                left = DataDistance.Pixels(value = 0),
-                top = DataDistance.Pixels(value = 139),
-                width = DataDistance.Pixels(value = 38),
-                align = null,
-                height = null,
-                topAnchor = null,
-                leftAnchor = null,
-                tooltip = "Jab Current Target",
-                value = "jab",
-                cmd = "jab",
-                echo = null,
-            ),
-            DialogObject.Button(
-                id = "cmdPunch",
-                left = DataDistance.Pixels(value = 38),
-                top = DataDistance.Pixels(value = 139),
-                width = DataDistance.Pixels(value = 53),
-                align = null,
-                height = null,
-                topAnchor = null,
-                leftAnchor = null,
-                tooltip = "Punch Current Target",
-                value = "punch",
-                cmd = "punch",
-                echo = null,
-            ),
-            DialogObject.Button(
-                id = "cmdGrapple",
-                left = DataDistance.Pixels(value = 91),
-                top = DataDistance.Pixels(value = 139),
-                width = DataDistance.Pixels(value = 59),
-                align = null,
-                height = null,
-                topAnchor = null,
-                leftAnchor = null,
-                tooltip = "Grapple Current Target",
-                value = "grapple",
-                cmd = "grapple",
-                echo = null,
-            ),
-            DialogObject.Button(
-                id = "cmdKick",
-                left = DataDistance.Pixels(value = 150),
-                top = DataDistance.Pixels(value = 139),
-                width = DataDistance.Pixels(value = 40),
-                height = null,
-                align = null,
-                topAnchor = null,
-                leftAnchor = null,
-                tooltip = "Kick Current Target",
-                value = "kick",
-                cmd = "kick",
-                echo = null,
-            ),
-            DialogObject.Button(
-                id = "cmdQuickstrike",
-                left = DataDistance.Pixels(value = 53),
-                top = DataDistance.Pixels(value = 165),
-                width = DataDistance.Pixels(value = 137),
-                height = DataDistance.Pixels(value = 20),
-                align = null,
-                topAnchor = null,
-                leftAnchor = null,
-                tooltip = "Quickstrike Next Action",
-                value = "prepare to quickstrike",
-                cmd = "quickstrike % uDEQuickstrike %",
-                echo = null,
-            ),
-            DialogObject.Link(
-                id = "lnSkin",
-                left = DataDistance.Pixels(value = 0),
-                top = DataDistance.Pixels(value = 191),
-                width = null,
-                height = null,
-                align = null,
-                topAnchor = null,
-                leftAnchor = null,
-                tooltip = null,
-                value = "skin",
-                cmd = "_skin",
-                echo = "skin",
-            ),
-            DialogObject.Link(
-                id = "lnSearch",
-                left = null,
-                top = null,
-                width = null,
-                height = null,
-                align = null,
-                topAnchor = null,
-                leftAnchor = null,
-                tooltip = null,
-                value = "search",
-                cmd = "_search",
-                echo = "search",
-            ),
-            DialogObject.Link(
-                id = "lnGrip",
-                left = null,
-                top = null,
-                width = null,
-                height = null,
-                align = null,
-                topAnchor = null,
-                leftAnchor = null,
-                tooltip = null,
-                value = "grip",
-                cmd = "_grip",
-                echo = "grip",
-            ),
-        )
-    DialogContent(
-        dataObjects = dialogData,
-        modifier = Modifier.size(190.dp, 219.dp),
-        executeCommand = {},
-        style = StyleDefinition(),
-    )
-}
-
-@Preview
-@Composable
-private fun ExperiencePreview() {
-    val dialogData =
-        listOf(
-            DialogObject.Label(
-                id = "yourLvl",
-                left = DataDistance.Pixels(value = 0),
-                top = DataDistance.Pixels(value = 0),
-                width = DataDistance.Pixels(value = 160),
-                height = DataDistance.Pixels(value = 15),
-                align = "n",
-                topAnchor = null,
-                leftAnchor = null,
-                tooltip = null,
-                value = "Level 0",
-            ),
-            DialogObject.ProgressBar(
-                id = "mindState",
-                left = DataDistance.Pixels(value = 3),
-                top = DataDistance.Pixels(value = 45),
-                width = DataDistance.Pixels(value = 160),
-                height = DataDistance.Pixels(value = 15),
-                align = "n",
-                topAnchor = null,
-                leftAnchor = null,
-                tooltip = null,
-                value = Percentage(value = 0),
-                text = "clear as a bell",
-            ),
-            DialogObject.ProgressBar(
-                id = "nextLvlPB",
-                left = DataDistance.Pixels(value = 3),
-                top = DataDistance.Pixels(value = 20),
-                width = DataDistance.Pixels(value = 160),
-                height = DataDistance.Pixels(value = 15),
-                align = "n",
-                topAnchor = null,
-                leftAnchor = null,
-                tooltip = null,
-                value = Percentage(value = 6),
-                text = "2365 until next level",
-            ),
-            DialogObject.Label(
-                id = "PTPs",
-                left = DataDistance.Pixels(value = 20),
-                top = DataDistance.Pixels(value = 0),
-                width = DataDistance.Pixels(value = 80),
-                height = DataDistance.Pixels(value = 20),
-                align = null,
-                topAnchor = "mindState",
-                leftAnchor = null,
-                tooltip = "Physical Training Points",
-                value = "5 PTPs",
-            ),
-            DialogObject.Label(
-                id = "MTPs",
-                left = DataDistance.Pixels(value = 0),
-                top = DataDistance.Pixels(value = 0),
-                width = DataDistance.Pixels(value = 80),
-                height = DataDistance.Pixels(value = 20),
-                align = null,
-                topAnchor = "mindState",
-                leftAnchor = "PTPs",
-                tooltip = "Mental Training Points",
-                value = "3 MTPs",
-            ),
-            DialogObject.Label(
-                id = "p2m",
-                left = DataDistance.Pixels(value = 20),
-                top = DataDistance.Pixels(value = 0),
-                width = DataDistance.Pixels(value = 80),
-                height = DataDistance.Pixels(value = 20),
-                align = null,
-                topAnchor = "PTPs",
-                leftAnchor = null,
-                tooltip = "Physical tps that have been converted to Mental tps",
-                value = "0 P2M",
-            ),
-            DialogObject.Label(
-                id = "m2p",
-                left = DataDistance.Pixels(value = 0),
-                top = DataDistance.Pixels(value = 0),
-                width = DataDistance.Pixels(value = 80),
-                height = DataDistance.Pixels(value = 20),
-                align = null,
-                topAnchor = "MTPs",
-                leftAnchor = "p2m",
-                tooltip = "Mental tps that have been converted to Physical tps",
-                value = "0 M2P",
-            ),
-            DialogObject.Link(
-                id = "exprLNK",
-                left = DataDistance.Pixels(value = 20),
-                top = DataDistance.Pixels(value = 0),
-                width = DataDistance.Pixels(value = 80),
-                height = null,
-                align = null,
-                topAnchor = "p2m",
-                leftAnchor = null,
-                tooltip = null,
-                value = "Details",
-                cmd = "experie",
-                echo = "experience",
-            ),
-            DialogObject.Link(
-                id = "goalsLNK",
-                left = DataDistance.Pixels(value = 0),
-                top = DataDistance.Pixels(value = 0),
-                width = DataDistance.Pixels(value = 80),
-                height = null,
-                align = null,
-                topAnchor = "m2p",
-                leftAnchor = "exprLNK",
-                tooltip = null,
-                value = "Skill Goals",
-                cmd = "goals",
-                echo = "goals",
-            ),
-        )
-    DialogContent(
-        dataObjects = dialogData,
-        modifier = Modifier.size(190.dp, 200.dp),
-        executeCommand = {},
-        style = StyleDefinition(),
-    )
-}
