+++ desktopApp/src/main/kotlin/warlockfe/warlock3/app/AppMenuBar.kt
@@ -4,67 +4,33 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.window.FrameWindowScope
 import androidx.compose.ui.window.MenuBar
 
-@Composable
-fun FrameWindowScope.AppMenuBar(
-    isConnected: Boolean,
-    runScript: () -> Unit,
-    openNewWindow: () -> Unit,
-    showSettingsDialog: () -> Unit,
-    showUpdateDialog: () -> Unit,
-    disconnect: () -> Unit,
-    showAboutDialog: () -> Unit,
-    exportSettings: () -> Unit,
-    exportCharacterSettings: (() -> Unit)?,
-    importSettings: () -> Unit,
-    importWraythSettings: () -> Unit,
-) {
-    MenuBar {
-        Menu("File") {
-            Item("New window", onClick = openNewWindow)
-            Item(
-                text = "Run script...",
-                enabled = isConnected,
-                onClick = runScript,
-            )
-            Separator()
-            Item(
-                text = "Export all settings...",
-                onClick = exportSettings,
-            )
-            Item(
-                text = "Export current character...",
-                enabled = exportCharacterSettings != null,
-                onClick = { exportCharacterSettings?.invoke() },
-            )
-            Item(
-                text = "Import settings...",
-                onClick = importSettings,
-            )
-            Item(
-                text = "Import wrayth settings...",
-                onClick = importWraythSettings,
-            )
-            Item(
-                text = "Settings",
-                onClick = showSettingsDialog,
-            )
-            Separator()
-            Item(
-                text = "Disconnect",
-                enabled = isConnected,
-                onClick = disconnect,
-            )
-        }
+// A platform-agnostic description of the app menus, defined once (see TitleBarView) and rendered as a
+// native menu bar on macOS ([AppMenuBar]) or as custom popup menus elsewhere, so the two renderers
+// can't drift. A null [AppMenu.items] entry is a separator.
+internal class AppMenu(
+    val title: String,
+    val items: List<AppMenuItem?>,
+)
 
-        Menu("Help") {
-            Item(
-                text = "Updates",
-                onClick = showUpdateDialog,
-            )
-            Item(
-                text = "About",
-                onClick = showAboutDialog,
-            )
+internal class AppMenuItem(
+    val label: String,
+    val enabled: Boolean = true,
+    val onClick: () -> Unit,
+)
+
+@Composable
+internal fun FrameWindowScope.AppMenuBar(menus: List<AppMenu>) {
+    MenuBar {
+        menus.forEach { menu ->
+            Menu(menu.title) {
+                menu.items.forEach { item ->
+                    if (item == null) {
+                        Separator()
+                    } else {
+                        Item(text = item.label, enabled = item.enabled, onClick = item.onClick)
+                    }
+                }
+            }
         }
     }
 }
